refactor(mitm-addon): localize webhook opener
